+++ ReleaseNotes/release_notes.md
@@ -36,7 +36,8 @@ minuit.release ('p3')
   1. add `**kwargs` for all `parallel`-methods, arguments are used for `WorkManager`
   1. remove `evaluateBatch` form all PDFs (folloiimng evolution of ROOT). We need to  gradually introduce `evaluateSpan`
   1. `Ostap::Math::GammaBW3`: use 1/s factor instead of 1/s^3/2. thanks to Misha Mikhasenko! 
-
+  1. Add symmetic   Sinh-Asinh resolution  model `ResoSinhAsinh`
+  1. Add symmetic Jonhson's SU resolution  model `ResoJohnsonSU`
 
 ## Backward incompatible changes: